[Doc] Remove bitness from pivx-qt help message and manpage
Remove the `(64-bit)` from the pivx-qt help message.

Since removing the Windows 32-bit builds, it is no longer information
that is often useful for troubleshooting. This never worked for other
architectures than x86, and the only 32-bit x86 build left is the Linux
one. Linux users tend to know what architecture they are using.

It also accidentally ends up in the pivx-qt manpage.