[GA] Fix GA for macOS runners
Homebrew no longer supports pre-built (binary) packages on macOS 11.
Some binary packages remain (for now). This means that packages are
built from source instead, which causes a runner timeout to occur when
trying to build the Qt and ZMQ packages.

A secondary issue came up with the newer Boost 1.83 version supplied by
homebrew for all macOS versions that was causing build errors.

To work around this, I've re-worked the CI and CMake workflows to
not include Qt/ZMQ on macOS 11 runners, and also set a specific Boost
version (1.76) that is still binary compatible with macOS 11